backend/requirements.txt
haystack-ai
python-dotenv
qdrant-haystack
docling
docling-haystack
google-genai-haystack
fastapi
uvicorn[standard]
python-multipart
